move _spark_script_args() and most of _interpolate_step_args() to base runner
_interpolate_jar_step_args() now handles GENERIC_ARGS
